Renamed types.py to otypes.py, to stop shadowing the python types module.
