+++ pyorient/__init__.py
@@ -2,7 +2,7 @@ __author__ = 'Ostico <ostico@gmail.com>'
 
 from .orient import OrientDB, OrientSocket
 from .exceptions import *
-from .types import *
+from .otypes import *
 from .constants import *
 from .scripts import Scripts
 from .serializations import OrientSerialization
